+++ pkg/consumergroup/exceptionconsumer/adapter.go
@@ -21,6 +21,8 @@ func toExceptionMessageMeta(exception *controller.ClaimedException, resolved *co
 		MessageKey:     exception.MessageKey,
 		CompactionRank: exception.CompactionRank,
 		CreatedAt:      exception.CreatedAt,
+		Attempts:       exception.Attempts,
+		Delays:         exception.Delays,
 		Options:        resolved,
 	}
 }